Remove timestamp from blob
I don't know what I was thinking - Git blobs don't have timestamps.  We
*might* do something in the future to try to figure out when a file
entered the repository/when the last change to a file was/whatever
(which would likely be expensive), but that's not now, and that's not
something we should act like we know now.  Out of all likelihood we're
just going to use the latest commit date